### 2019-10-02 #1779 @whedon
[PRE REVIEW]: Telewavesim: Python software for teleseismic body wave modeling
**Submitting author:** @paudetseis (<a href="http://orcid.org/0000-0003-2364-9454">Pascal Audet</a>)
**Repository:** <a href="https://github.com/paudetseis/Telewavesim" target ="_blank">https://github.com/paudetseis/Telewavesim</a>
**Version:** v0.1.0
