### 2019-10-07 7259bc6
ci: add .travis.yml file
