### 2019-10-07 69c4ac1
ci: add fortran compile
see https://travis-ci.org/trichter/Telewavesim/builds/594466321
