+++ .travis.yml
@@ -18,9 +18,14 @@ matrix:
 before_install:
   - wget https://raw.githubusercontent.com/trichter/conda4travis/latest/conda4travis.sh -O conda4travis.sh
   - source conda4travis.sh
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
+      GF="gfortran_osx-64";
+    elif [[ "$TRAVIS_OS_NAME" == "linux" ]]; then
+      GF="gfortran_linux-64";
+    fi
   - conda config --add channels conda-forge
   - conda create -q -n testenv
-      python=$PYTHON_VERSION obspy pyfftw
+      python=$PYTHON_VERSION obspy pyfftw $GF
       pytest-cov
   - conda activate testenv
   - conda list
