### 2019-10-07 600bbae
apply patch from #2 to get it working on linux
see https://travis-ci.org/trichter/Telewavesim/builds/594469551
