### 2019-10-07 34a32c4
only apply patch on linux
