### 2019-10-07 63af1cc
make patch stable
