### 2019-10-07 0243420
ci: explicitely install lapack on linux
