### 2019-10-07 a0e28cb
ci: exclude linux python 3.5 from build matrix
