+++ .travis.yml
@@ -3,8 +3,6 @@ sudo: false
 matrix:
   include:
     - os: linux
-      env: PYTHON_VERSION=3.5
-    - os: linux
       env: PYTHON_VERSION=3.6
     - os: linux
       env: PYTHON_VERSION=3.7
