### 2019-10-07 a0d6c6c
make patch even more stable
* necessary because np.show_config() sometimes returns None
