+++ setup.py
@@ -5,9 +5,14 @@ if platform.system() == 'Linux':
     # this hack hopefully  will become unecessary in the future
     # see issue #2
     # ci fail: https://travis-ci.org/trichter/Telewavesim/builds/594469551
-    from numpy import show_config
-    libdirs = {c['library_dirs'] for c in show_config().items()
-               if c is not None and 'library_dirs' in c}
+    from numpy import __config__
+    attrs = ['blas_mkl_info', 'blas_opt_info', 'lapack_mkl_info',
+             'lapack_opt_info', 'openblas_lapack_info']
+    libdirs = set()
+    for attr in attrs:
+        obj = getattr(__config__, attr, {}).get('library_dirs')
+        if obj is not None:
+            libdirs.update(obj)
     ext_kw = dict(library_dirs=list(libdirs), libraries=['lapack'])
 else:
     ext_kw = {}
