### 2019-10-07 97cdd1b
ci: fix coverage
