### 2019-10-07 d7338b8
move tests inside package to provide them on pypi
