### 2019-10-07 ab328aa
tests: discover model file dynamically
