+++ telewavesim/tests/conftest.py
@@ -1,23 +1,21 @@
 import pytest
 from telewavesim import conf as cf
 from telewavesim import utils as ut
+from pkg_resources import resource_filename
+
 
 @pytest.fixture(scope="module")
 def load_params():
-
-    try:
-        modfile='tests/test_model_Audet2016.txt'
-        ut.read_model(modfile)
-    except:
-        modfile='test_model_Audet2016.txt'
-        ut.read_model(modfile)
+    modfile = resource_filename('telewavesim',
+                                'examples/models/model_Audet2016.txt')
+    ut.read_model(modfile)
 
     cf.nt = 2000
     cf.dt = 0.05
     cf.slow = 0.04
     cf.baz = 0.
     cf.dp = 1000.
-    cf.wvtype='P'
+    cf.wvtype = 'P'
 
     ut.model2for()
     ut.wave2for()
