### 2019-10-07 a6f759d
add notebook codes as test
