### 2019-10-07 19c8fa7
remove conf import in wiggle module
