+++ telewavesim/tests/test_notebooks.py
@@ -63,7 +63,8 @@ def test_Porter2011():
     with tempfile.TemporaryDirectory() as tempdir:
         wg.rf_wiggles_baz(trR, trT, trR_stack, trT_stack, 'test', btyp='baz',
                           scale=1.e3, tmin=-5., tmax=8., save=True,
-                          ftitle=join(tempdir, 'porter2011.png'))
+                          ftitle=join(tempdir, 'porter2011.png'),
+                          wvtype='P')
 
 
 def test_Autdet2016():
@@ -152,4 +153,5 @@ def test_SKS():
     with tempfile.TemporaryDirectory() as tempdir:
         wg.pw_wiggles_baz(trR, trT, 'test', btyp='baz', scale=0.05,
                           t1=t1, tmin=0., tmax=40, save=True,
-                          ftitle=join(tempdir, 'sks.png'))
+                          ftitle=join(tempdir, 'sks.png'),
+                          wvtype='SV')
