### 2019-10-07 7af55b1
utils: remove unused statements
