+++ telewavesim/utils.py
@@ -25,14 +25,13 @@
 Utility functions to interact with ``telewavesim`` modules.
 
 '''
-import sys
 import itertools
 import numpy as np
 import pyfftw
 from scipy.signal import hilbert
 from obspy.core import Trace, Stream
 from obspy.signal.rotate import rotate_ne_rt
-from telewavesim import conf as cf 
+from telewavesim import conf as cf
 from telewavesim import elast as es
 from telewavesim.rmat_f import conf as cf_f
 from telewavesim.rmat_f import plane as pw_f
@@ -40,7 +39,7 @@ from telewavesim.rmat_f import plane as pw_f
 
 def set_iso_tensor(a, b):
     """
-    Function to generate tensor for isotropic material. 
+    Function to generate tensor for isotropic material.
 
     Args:
         a (float): P-wave velocity (km/s)
@@ -73,7 +72,7 @@ def set_tri_tensor(a, b, tr, pl, ani):
         b (float): S-wave velocity (km/s)
         tr (float): Trend angle of symmetry axis (degree)
         pl (float): Plunge angle of symmetry axis (degree)
-        ani (float): Percent anisotropy 
+        ani (float): Percent anisotropy
 
     Returns:
         (np.ndarray): cc: Elastic tensor (GPa /density) \
@@ -88,14 +87,14 @@ def set_tri_tensor(a, b, tr, pl, ani):
     # Percent anisotropy
     da = (a*1.e3)*ani/100.
     db = (b*1.e3)*ani/100.
-    
+
     # Set up matrix elements
     AA = (a*1.e3 - da/2.)**2
     CC = (a*1.e3 + da/2.)**2
     LL = (b*1.e3 + db/2.)**2
-    NN = (b*1.e3 - db/2.)**2 
+    NN = (b*1.e3 - db/2.)**2
     AC = (a*1.e3)**2
-    FF = -LL + np.sqrt((2.*AC)**2 - 2.*AC*(AA + CC + 2.*LL) + 
+    FF = -LL + np.sqrt((2.*AC)**2 - 2.*AC*(AA + CC + 2.*LL) +
             (AA + LL)*(CC + LL))
     # eta = FF/(AA - 2.*LL)
 
@@ -223,7 +222,7 @@ def voigt2cc(C):
         (np.ndarray): cc: Elastic tensor (shape ``(3, 3, 3, 3)``)
 
     """
-    
+
     C = np.asarray(C)
     cc = np.zeros((3,3,3,3), dtype=float)
     for i, j, k, l in itertools.product(range(3), range(3), range(3), range(3)):
@@ -236,7 +235,7 @@ def voigt2cc(C):
 def cc2voigt(cc):
     """
     Convert from the full 3x3x3x3 tensor representation
-    to the Voigt notation of the stiffness matrix. 
+    to the Voigt notation of the stiffness matrix.
 
     Args:
         cc (np.ndarray): Elastic tensor (shape ``(3, 3, 3, 3)``)
@@ -247,8 +246,6 @@ def cc2voigt(cc):
     """
 
     Voigt_notation = [(0, 0), (1, 1), (2, 2), (1, 2), (0, 2), (0, 1)]
-
-    tol = 1e-3
 
     cc = np.asarray(cc)
     C = np.zeros((6,6))
@@ -270,8 +267,8 @@ def VRH_average(C):
     Args:
         C (np.ndarray): Stiffness matrix (shape ``(6, 6)``)
 
-    Returns:     
-        (tuple): Tuple containing: 
+    Returns:
+        (tuple): Tuple containing:
             * Kvoigt (float): Voigt average bulk modulus (GPa)
[…]
