### 2019-10-08 7144619
WIP: remove conf module and use arguments instead
* finished, but a bug was introduced which I was not able to identify
* run pytest to see the bug
