+++ telewavesim/utils.py
@@ -1,4 +1,4 @@
-# Copyright 2019 Pascal Audet
+# Copyright 2019 Pascal Audet, Tom Eulenfeld
 
 # This file is part of Telewavesim.
 
@@ -31,7 +31,6 @@ import pyfftw
 from scipy.signal import hilbert
 from obspy.core import Trace, Stream
 from obspy.signal.rotate import rotate_ne_rt
-from telewavesim import conf as cf
 from telewavesim import elast as es
 from telewavesim.rmat_f import conf as cf_f
 from telewavesim.rmat_f import plane as pw_f
@@ -406,7 +405,7 @@ def rot_tensor(a,alpha,beta,gam):
     return aa
 
 
-def rotate_zrt_pvh(trZ, trR, trT, vp=6., vs=3.5):
+def rotate_zrt_pvh(trZ, trR, trT, slow, vp=6., vs=3.5):
     """
     Rotates traces from `Z-R-T` orientation to `P-SV-SH` wave mode.
 
@@ -414,6 +413,7 @@ def rotate_zrt_pvh(trZ, trR, trT, vp=6., vs=3.5):
         trZ (obspy.trace): Vertical component
         trR (obspy.trace): Radial component
         trT (obspy.trace): Transverse component
+        slow (float): slowness of wave
         vp (float, optional): P-wave velocity used for rotation
         vs (float, optional): S-wave velocity used for rotation
 
@@ -431,14 +431,14 @@ def rotate_zrt_pvh(trZ, trR, trT, vp=6., vs=3.5):
     trH = trT.copy()
 
     # Vertical slownesses
-    qp = np.sqrt(1/vp/vp - cf.slow*cf.slow)
-    qs = np.sqrt(1/vs/vs - cf.slow*cf.slow)
+    qp = np.sqrt(1/vp**2 - slow**2)
+    qs = np.sqrt(1/vs**2 - slow**2)
 
     # Elements of rotation matrix
-    m11 = cf.slow*vs*vs/vp
-    m12 = -(1 - 2*vs*vs*cf.slow*cf.slow)/(2*vp*qp)
-    m21 = (1 - 2*vs*vs*cf.slow*cf.slow)/(2*vs*qs)
-    m22 = cf.slow*vs
+    m11 = slow*vs*vs/vp
+    m12 = -(1 - 2*vs*vs*slow*slow)/(2*vp*qp)
+    m21 = (1 - 2*vs*vs*slow*slow)/(2*vs*qs)
+    m22 = slow*vs
 
     # Rotation matrix
     rot = np.array([[-m11, m12], [-m21, m22]])
@@ -520,7 +520,7 @@ def stack_all(st1, st2, pws=False):
     return stack1, stack2
 
 
-def calc_ttime(slow):
+def calc_ttime(model, slow, wvtype='P'):
     """
     Calculates total propagation time through model. The
     bottom layer is irrelevant in this calculation.
@@ -565,118 +565,115 @@ def calc_ttime(slow):
 
     t1 = 0.
 
-    for i in range(cf.nlay-1):
-        if cf.isoflg[i] == 'iso':
-            a0 = cf.a[2,2,2,2,i]
-            b0 = cf.a[1,2,1,2,i]
+    for i in range(model.nlay-1):
+        if model.isoflg[i] == 'iso':
+            a0 = model.a[2,2,2,2,i]
+            b0 = model.a[1,2,1,2,i]
         else:
-            cc = cc2voigt(cf.a[:,:,:,:,i])
-            rho = cf.rho[i]
+            cc = cc2voigt(model.a[:,:,:,:,i])
+            rho = model.rho[i]
[…]
