+++ telewavesim/tests/test_notebooks.py
@@ -31,24 +31,21 @@ def test_Porter2011():
     from obspy.core import Stream
     from telewavesim import utils as ut
     from telewavesim import wiggle as wg
-    from telewavesim import conf as cf
 
     modfile = resource_filename('telewavesim',
                                 'examples/models/model_Porter2011.txt')
-    cf.wvtype = 'P'
-    cf.nt = 3000  # Number of samples
-    cf.dt = 0.01  # Sample distance in seconds
-    cf.slow = 0.06  # Horizontal slowness (or ray parameter) in s/km
+    wvtype = 'P'
+    npts = 3000  # Number of samples
+    dt = 0.01  # Sample distance in seconds
+    slow = 0.06  # Horizontal slowness (or ray parameter) in s/km
     baz = np.arange(0., 360., 10.)
-    ut.read_model(modfile)
+    model = ut.read_model(modfile)
     trR = Stream()
     trT = Stream()
     # Loop over range of data
     for bb in baz:
-        # Pass baz global variable
-        cf.baz = bb
         # Calculate the plane waves seismograms
-        trxyz = ut.run_plane()
+        trxyz = ut.run_plane(model, slow, npts, dt, bb, wvtype=wvtype, obs=False)
         # Then the transfer functions in Z-R-T coordinate system
         tfs = ut.tf_from_xyz(trxyz, pvh=False)
         # Append to streams
@@ -74,30 +71,30 @@ def test_Autdet2016():
     from obspy.signal.rotate import rotate_ne_rt
     from telewavesim import utils as ut
     from telewavesim import wiggle as wg
-    from telewavesim import conf as cf
     modfile = resource_filename('telewavesim',
                                 'examples/models/model_Audet2016.txt')
-    cf.wvtype = 'P'
-    cf.nt = 3000  # Number of samples
-    cf.dt = 0.01  # Sample distance in seconds
-    cf.dp = 2000.  # Deployment depth below sea level in meters
-    cf.c = 1500.    # P-wave velocity in salt water (m/s)
-    cf.rhof = 1027.  # Density of salt water (kg/m^3)
-    cf.slow = 0.06  # Horizontal slowness (or ray parameter) in s/km
+    wvtype = 'P'
+    npts = 3000  # Number of samples
+    dt = 0.01  # Sample distance in seconds
+    dp = 2000.  # Deployment depth below sea level in meters
+    c = 1.500    # P-wave velocity in salt water (km/s)
+    rhof = 1027.  # Density of salt water (kg/m^3)
+    slow = 0.06  # Horizontal slowness (or ray parameter) in s/km
     # Back-azimuth direction in degrees (has no influence if model is isotropic)
-    cf.baz = 0.
-    ut.read_model(modfile)
-    assert cf.rho == [2800.0, 2800.0, 3200.0]
-    t1 = ut.calc_ttime(cf.slow)
+    baz = 0.
+    model = ut.read_model(modfile)
+    assert model.rho == [2800.0, 2800.0, 3200.0]
+    t1 = ut.calc_ttime(model, slow, wvtype=wvtype)
     assert round(t1, 1) == 1.1
-    trxyz = ut.run_plane(obs=True)
+    trxyz = ut.run_plane(model, slow, npts, dt, baz=baz, wvtype=wvtype,
+                         obs=True, dp=dp, c=c, rhof=rhof)
     tfs = ut.tf_from_xyz(trxyz, pvh=False)
     ntr = trxyz[0]  # North component
     etr = trxyz[1]  # East component
     ztr = trxyz[2]  # Vertical component
     rtr = ntr.copy()  # Radial component
     ttr = etr.copy()  # Transverse component
[…]
