+++ telewavesim/tests/conftest.py
@@ -8,17 +8,19 @@ from pkg_resources import resource_filename
 def load_params():
     modfile = resource_filename('telewavesim',
                                 'examples/models/model_Audet2016.txt')
-    ut.read_model(modfile)
+    model = ut.read_model(modfile)
 
     cf.nt = 2000
-    cf.dt = 0.05
-    cf.slow = 0.04
-    cf.baz = 0.
-    cf.dp = 1000.
     cf.wvtype = 'P'
+    dt = 0.05
+    slow = 0.04
+    baz = 0.
+    dp = 1000
+    c = 1.5
+    rhof = 1027
 
-    ut.model2for()
-    ut.wave2for()
-    ut.obs2for()
+    model2for(model)
+    wave2for(dt, slow, baz)
+    obs2for(dp, c, rhof)
 
     return

[…]
