### 2019-10-08 57b0645
convert model inputs to ndarray
