### 2019-10-08 aaa5675
simplify calc_ttime example, make it work again and correct output
