### 2019-10-08 49fb1c0
update doc in Model and calc_ttime
