### 2019-10-08 3eeb3f7
eliminate potential bug when calling subroutines multiple times
