### 2019-10-08 57961b8
update doc in utils module
