### 2019-10-08 585151d
finally remove conf and tests.conftest modules
