### 2019-10-08 0e943c3
readme: update installation instructions
