### 2019-10-08 #1779 @paudetseis
Hi @Kevin-Mattheus-Moerman, @andreww and @seisman
Thanks for handling the paper. I'd just like to point out that I just merged a PR from @trichter that improved the code significantly. You might want to do a git pull if you already cloned the repo.
