### 2019-10-08 fa97ded
removed conf from software
