### 2019-10-09 9cceeab
updated installation instructions
