+++ docs/_sources/elast.rst.txt
@@ -1,8 +1,8 @@
 .. figure:: ../telewavesim/examples/picture/tws_logo.png
    :align: center
 
-telewavesim.elast
------------------
+Module elast
+------------
 
 .. automodule:: telewavesim.elast
    :members:
