+++ docs/_sources/wiggles.rst.txt
@@ -1,8 +1,8 @@
 .. figure:: ../telewavesim/examples/picture/tws_logo.png
    :align: center
 
-telewavesim.wiggles
--------------------
+Module wiggles
+--------------
 
 .. automodule:: telewavesim.wiggle
    :members:
