### 2019-10-15 6768b1e
add telewavesim.__version__ string, single place to set the version number
