+++ telewavesim/tests/test_notebooks.py
@@ -20,7 +20,6 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-
 from pkg_resources import resource_filename
 import tempfile
 from os.path import join
