+++ .travis.yml
@@ -21,7 +21,7 @@ before_install:
     fi
   - conda config --add channels conda-forge
   - conda create -q -n testenv
-      python=$PYTHON_VERSION obspy pyfftw $GF $LP
+      python=$PYTHON_VERSION obspy $GF $LP
       pytest-cov
   - conda activate testenv
   - conda list
