### 2019-10-16 c4bd726
Updated docs to reflect new installation instructions
