+++ docs/_sources/conf.rst.txt
@@ -1,8 +1,0 @@
-.. figure:: ../telewavesim/examples/picture/tws_logo.png
-   :align: center
-
-telewavesim.conf
-----------------
-
-.. automodule:: telewavesim.conf
-   :members:
