+++ docs/_sources/elast.rst.txt
@@ -1,8 +1,0 @@
-.. figure:: ../telewavesim/examples/picture/tws_logo.png
-   :align: center
-
-Module elast
-------------
-
-.. automodule:: telewavesim.elast
-   :members:
