+++ docs/_sources/api.rst.txt
@@ -1,5 +1,0 @@
-.. figure:: ../telewavesim/examples/picture/tws_logo.png
-   :align: center
-
-.. automodule:: telewavesim
-   :members:
