+++ docs/_sources/rmat_f.rst.txt
@@ -1,8 +1,0 @@
-.. figure:: ../telewavesim/examples/picture/tws_logo.png
-   :align: center
-
-telewavesim.rmat_f
-------------------
-
-.. automodule:: telewavesim.rmat_f
-   :members:
