### 2019-10-16 #1779 @kbarnhart
@paudetseis I'd like to find one more reviewer for your submission. If you have any recommendations, please mention them on this thread. The following is a [list of potential reviewers](https://bit.ly/joss-reviewers). You can also recommend people who are not on this list.
