### 2019-10-16 #1779 @kbarnhart
@paudetseis thanks for that recommendation.

@ThomasLecocq and @brmather would you be interested in reviewing this work for the [Journal of Open Source Software (JOSS)](https://joss.theoj.org)? [**Telewavesim: Software for teleseismic body wave modeling**](https://github.com/paudetseis/Telewavesim)

The streamlined [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) takes place on GitHub and focuses on the software and a [very short paper](https://github.com/openjournals/joss-papers/blob/joss.01779/joss.01779/10.21105.joss.01779.pdf).

This here is a pre-review issue. I would like one more reviewer and then I'll open a dedicated review issue where check-boxes guide you through the review process.

Let me know if you have questions.
