### 2019-10-17 #1779 @brmather
Hi @kbarnhart and @paudetseis, I would be happy to review [Telewavesim](https://github.com/paudetseis/Telewavesim).for JOSS.
