### 2019-10-18 #1818 @whedon
[REVIEW]: Telewavesim: Python software for teleseismic body wave modeling
**Submitting author:** @paudetseis (<a href="http://orcid.org/0000-0003-2364-9454">Pascal Audet</a>)
**Repository:** <a href="https://github.com/paudetseis/Telewavesim" target ="_blank">https://github.com/paudetseis/Telewavesim</a>
**Version:** v0.2.0
