### 2019-10-24 paudetseis/Telewavesim#10 @seisman
[JOSS review] Supported Python versions
It's not clear to me which Python versions are really supported.

- The README says `The current version was developed using Python3.7`. It sounds like Python 3.7 is the only officially supported version for me.
- The travis CI tests run well for 3.6 and 3.7
- The setup.py says it works for Python>=3.5.

If the package works well for Python>=3.5, it's better to test it with Py3.5.

https://github.com/openjournals/joss-reviews/issues/1818
