### 2019-10-24 paudetseis/Telewavesim#11 @seisman
[JOSS review] version.txt is not used
It seems the package version information is written in the [\_\_init__.py](https://github.com/paudetseis/Telewavesim/blob/master/telewavesim/__init__.py) and automatically detected in [setup.py](https://github.com/paudetseis/Telewavesim/blob/master/setup.py).

`version.txt` isn't used anywhere and should be safe to be deleted.

https://github.com/openjournals/joss-reviews/issues/1818
