### 2019-10-24 paudetseis/Telewavesim#11 @paudetseis
Thanks @seisman - I am making these corrections as you are sending them. Should I push them now or wait for potentially more comments from you or other reviewers?
