### 2019-10-24 #1818 @paudetseis
Question to @kbarnhart and reviewers: I see new issues on the GH repo from @seisman and making corrections as they come. Should I push them now or wait for potentially more comments from reviewers?
