### 2019-10-25 paudetseis/Telewavesim#12 @seisman
[JOSS review] Python codes should follow pep8
Although not mandatory, it's a good practice to follow the Python style guide [pep8](https://www.python.org/dev/peps/pep-0008/).

There are some tools to automatically format Python codes, e.g. [pycodestyle](https://github.com/PyCQA/pycodestyle), [black](https://github.com/psf/black).

openjournals/joss-reviews#1818
