### 2019-10-25 paudetseis/Telewavesim#10 @trichter
The package works well for Python>=3.5.

I removed the tests for python 3.5 at some point, because an old numpy version was installed along side obspy, see https://travis-ci.org/paudetseis/Telewavesim/builds/595021563.

I will try to add python 3.5 to the test matrix, again.
