### 2019-10-25 d207e97
ci: add py3.5 to the build matrix
