+++ .travis.yml
@@ -3,9 +3,13 @@ sudo: false
 matrix:
   include:
     - os: linux
+      env: PYTHON_VERSION=3.5
+    - os: linux
       env: PYTHON_VERSION=3.6
     - os: linux
       env: PYTHON_VERSION=3.7
+    - os: osx
+      env: PYTHON_VERSION=3.5
     - os: osx
       env: PYTHON_VERSION=3.6
     - os: osx
