### 2019-10-25 715879d
remove lapack work-around
