### 2019-10-25 dacaa5c
ci: add NPY_DISTUTILS_APPEND_FLAGS
