### 2019-10-25 da938f5
ci: specify minimal numpy version
