### 2019-10-25 76c49cb
ci: refactor .travis.yml
