### 2019-10-25 29c1152
pep8 in setup.py
