### 2019-10-25 53d54cc
tests: use agg mpl backend
