### 2019-10-25 867b25b
pep8 in test_notebooks
