### 2019-10-25 #1818 @kbarnhart
@paudetseis - it’s up to you if you want to wait for all the reviews, or start addressing the issues as they arrive.

It’s definitely good to do as @seisman has done and make sure to to link the reviews in the repository. That way the substantive review comments can be in the in-repository issues, but this issue can have a record of the the status.
