### 2019-10-25 e3f75e8
updated installation and usage documentation
