### 2019-10-26 paudetseis/Telewavesim#15 @andreww
[JOSS review] references in documentation
I see there are several citations in the documentation, for example "from Aleksandrov and Ryzhova (1986)" as the source of the [elastic constants data for biotite](https://paudetseis.github.io/Telewavesim/elast.html#telewavesim.elast.biotite). However, I cannot see an obvious reference list or very easily find the source. I'm not sure what the best way to do it is, but it would be nice to find some way of including the full reference (and ideally a link via the DOI).

  openjournals/joss-reviews#1818
