### 2019-10-26 7d3b875
removed version.txt and tws_env.yml
