+++ tws_env.yml
@@ -1,67 +1,0 @@
-name: tws
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - asn1crypto=0.24.0=py37_1003
-  - basemap=1.2.0=py37hd103db2_4
-  - bzip2=1.0.8=h01d97ff_1
-  - ca-certificates=2019.9.11=hecc5488_0
-  - certifi=2019.9.11=py37_0
-  - cffi=1.12.3=py37hccf1714_0
-  - chardet=3.0.4=py37_1003
-  - cryptography=2.7=py37h212c5bf_0
-  - cycler=0.10.0=py_1
-  - decorator=4.4.0=py_0
-  - freetype=2.10.0=h24853df_1
-  - future=0.17.1=py37_1000
-  - geos=3.7.2=h6de7cb9_2
-  - icu=64.2=h6de7cb9_1
-  - idna=2.8=py37_1000
-  - kiwisolver=1.1.0=py37h770b8ee_0
-  - libblas=3.8.0=12_openblas
-  - libcblas=3.8.0=12_openblas
-  - libcxx=9.0.0=0
-  - libcxxabi=9.0.0=0
-  - libffi=3.2.1=h6de7cb9_1006
-  - libgfortran=4.0.0=2
-  - libiconv=1.15=h01d97ff_1005
-  - liblapack=3.8.0=12_openblas
-  - libopenblas=0.3.7=h4bb4525_1
-  - libpng=1.6.37=h2573ce8_0
-  - libxml2=2.9.9=h12c6b28_5
-  - libxslt=1.1.33=h320ff13_0
-  - llvm-openmp=9.0.0=h40edb58_0
-  - lxml=4.4.1=py37h08abf6f_0
-  - matplotlib=3.1.1=py37_1
-  - matplotlib-base=3.1.1=py37h3a684a6_1
-  - ncurses=6.1=h0a44026_1002
-  - numpy=1.15.4=py37hbb3c62a_1002
-  - obspy=1.1.1=py37h917ab60_1
-  - openssl=1.1.1c=h01d97ff_0
-  - pip=19.2.3=py37_0
-  - proj4=4.9.3=h01d97ff_9
-  - pycparser=2.19=py37_1
-  - pyfftw=0.11.1=py37h917ab60_1001
-  - pyopenssl=19.0.0=py37_0
-  - pyparsing=2.4.2=py_0
-  - pyproj=1.9.5.1=py37hfb6cdf6_1006
-  - pyshp=2.1.0=py_0
-  - pysocks=1.7.1=py37_0
-  - python=3.7.3=h93065d6_1
-  - python-dateutil=2.8.0=py_0
-  - readline=8.0=hcfe32e1_0
-  - requests=2.22.0=py37_1
-  - scipy=1.3.1=py37h7e0e109_2
-  - setuptools=41.2.0=py37_0
-  - six=1.12.0=py37_1000
-  - sqlalchemy=1.3.8=py37h01d97ff_0
-  - sqlite=3.29.0=hb7d70f7_1
-  - tk=8.6.9=h2573ce8_1003
-  - tornado=6.0.3=py37h01d97ff_0
-  - urllib3=1.25.6=py37_0
-  - wheel=0.33.6=py37_0
-  - xz=5.2.4=h1de35cc_1001
-  - zlib=1.2.11=h01d97ff_1006
-prefix: /Users/pascalaudet/anaconda3/envs/tws
-
