### 2019-10-26 f07ea63
added contribution subsection in README
