### 2019-10-26 paudetseis/Telewavesim#16 @andreww
[JOSS review] telewavesim.elast.tri_tensor is the odd one out
The function `telewavesim.elast.tri_tensor` returns a 3x3x3x3 numpy array representing  the elasticity as a rank four tensor. All the other functions in this module (e.g. `telewavesim.elast.iso_tensor` return a 6x6 numpy array representing the elasticity in Voigt notation. It would be more consistent to return the elasticity in the same representation.

Also, what units should be used for density when normalising the elastic stiffnesses in GPa?

openjournals/joss-reviews#1818
