### 2019-10-26 ea3bd6b
clean up docs in utils
