### 2019-10-26 paudetseis/Telewavesim#17 @andreww
[JOSS review] What does utils.calc_ttime do?
As far as I can see, I'm supposed to be able to set up a generally anisotropic model and use `calc_ttime` to calculates total propagation time for waves of different types. However, I will get the same answer for SV and SH waves. I think a fuller explanation of what ttime is supposed to do would be useful.

openjournals/joss-reviews#1818
