### 2019-10-26 paudetseis/Telewavesim#18 @andreww
[JOSS review] No community guidelines
As far as I can see, there are no "community guidelines", which are requested by JOSS. They say:

> There should be clear guidelines for third-parties wishing to:
>
> Contribute to the software
> Report issues or problems with the software
> Seek support

This would be something like the "contributing" section in the readme for https://github.com/kb-press/ndsplines/, along with some information on how to get help.

openjournals/joss-reviews#1818
