### 2019-10-26 paudetseis/Telewavesim#19 @andreww
[JOSS review] Documentation improvements 
Overall, I think the documentation is OK for JOSS, but I think something is missing that would make it easer for a new user to make use of the software. What I think is missing is the information needed by somebody who wants to use the functions to do something different from the calculations presented in the three example notebooks. What I think is needed to allow this is a bit more descriptive information about what the key functions *do* alongside the basic API information (input and output arguments) and some brief 'primer' material to help people put this into context.

For example, `utils.run_plane` is a "function to run the plane module", but the documentation does not say what the plane module / model is. This doesn't need to be a lot of text, but some figures would probably help.

openjournals/joss-reviews#1818
