### 2019-10-26 paudetseis/Telewavesim#20 @andreww
[JOSS review] paper
The paper looks fine to me. Just a couple of minor things I spotted:

- [x] the brackets around "2017" are not needed for the Brownlee et al. citation in the second paragraph.
- [x] there is a spurious '}' after the Frazer & Gettrust citation in the third paragraph.
- [x] the reference for Kennett (1983) is incomplete - please include the book title.

openjournals/joss-reviews#1818
