### 2019-10-26 paudetseis/Telewavesim#21 @andreww
[JOSS review] no generally anisotropic example
A key claim in the paper is that the code handles general anisotropy and it looks like this is the case. However, the the layers in the three examples are either isotropic or have only transverse anisotropy. An example demonstrating general anisotropy would be useful in demonstrating the functionality of the software.

openjournals/joss-reviews#1818
