### 2019-10-26 #1818 @andreww
Ok - my review is done. Comments are all added as issues in the telewavesim repository and reference this issue. As I see things, the blockers would be paudetseis/Telewavesim#21 (functionality) and paudetseis/Telewavesim#18 (community guidelines).
