### 2019-10-28 5dd2d24
revised contributions section in README
