### 2019-10-28 b92d1c3
added documentation for run_plane function with examples
