+++ telewavesim/utils.py
@@ -98,7 +98,35 @@ def set_tri_tensor(a, b, tr, pl, ani):
     # eta = FF/(AA - 2.*LL)
 
     # Get tensor with horizontal axis
-    cc = es.tri_tensor(AA, CC, FF, LL, NN)
+    cc = np.zeros((3,3,3,3))
+        
+    cc[0,0,0,0] = AA
+    cc[1,1,1,1] = AA
+    cc[2,2,2,2] = CC
+
+    cc[0,0,1,1] = (AA - 2.*NN)
+    cc[1,1,0,0] = (AA - 2.*NN)
+
+    cc[0,0,2,2] = FF
+    cc[2,2,0,0] = FF
+
+    cc[1,1,2,2] = FF
+    cc[2,2,1,1] = FF
+
+    cc[1,2,1,2] = LL
+    cc[2,1,2,1] = LL
+    cc[2,1,1,2] = LL
+    cc[1,2,2,1] = LL
+
+    cc[2,0,2,0] = LL
+    cc[0,2,0,2] = LL
+    cc[2,0,0,2] = LL
+    cc[0,2,2,0] = LL
+ 
+    cc[0,1,0,1] = NN
+    cc[1,0,1,0] = NN
+    cc[0,1,1,0] = NN
+    cc[1,0,0,1] = NN
 
     # Rotate tensor using trend and plunge
     cc = rot_tensor(cc, pl, tr, 0.)
@@ -522,7 +550,8 @@ def stack_all(st1, st2, pws=False):
 
 def calc_ttime(model, slow, wvtype='P'):
     """
-    Calculates total propagation time through model. The
+    Calculates total propagation time through model given the corresponding
+    wave type. The
     bottom layer is irrelevant in this calculation.
 
     .. note::
@@ -543,7 +572,7 @@ def calc_ttime(model, slow, wvtype='P'):
     Example
     -------
     >>> from telewavesim import utils
-    >>> # Define two-layer model model with identical material
+    >>> # Define two-layer model with identical material
     >>> model = utils.Model([10, 0], None, 0, 0, 'atg', 0, 0, 0)
     >>> # Only topmost layer is useful for travel time calculation
     >>> wvtype = 'P'
@@ -702,16 +731,32 @@ def obs2for(dp, c, rhof):
     cf_f.rhof = rhof
 
 
-def run_plane(model, slow, npts, dt, baz=0, wvtype='P', obs=False, dp=None, c=1.5, rhof=1027):
+def run_plane(model, slow, npts, dt, baz=0, wvtype='P', obs=False, dp=100., c=1.5, rhof=1027):
     """
     Function to run the ``plane`` module and return 3-component seismograms as an ``obspy``
-    ``Stream`` object.
+    ``Stream`` object. This function builds the seismic response spectrum frequency by frequency using the
+    matrix propagation approach. Required arguments are the seismic model, slowness value, and the sampling
+    properties (maximum number of samples and the sampling distance in seconds). 
+
+    By default, the function uses
+    a back-azimuth value of 0 degree, which is suitable for events coming from the North pole or isotropic 
+    seismic velocity models (i.e., those that do not vary with direction of incoming waves). For anisotropic
+    velocity models, users need to specify the back-aZimiuth value in degrees. Furthermore, the default
+    type of the incoming teleseismic body wave is ``'P'`` for compressional wave. Other options are 
+    ``'SV'``, ``'SH'``, or ``'Si'`` for vertically-polarized shear wave, horizontally-polarized shear wave
+    or isotropic shear wave, respectively. Wave modes cannot be mixed. 
+
[…]
