### 2019-10-28 c7c5133
fixed Issue #20 in JOSS paper
