+++ paper/paper.bib
@@ -55,7 +55,7 @@
 
 @book{Kennett:1983,
     author = {Kennett, B L N},
-    booktitle = {Seismic wave propagation in stratified media},
+    title = {Seismic wave propagation in stratified media},
     publisher = {Cambridge University Press},
     year = {1983}
 }
