### 2019-10-28 8af7cc6
Better documentation for function utils.calc_ttime from Issue #17 in JOSS review
