### 2019-10-28 3935dfe
added reference to Browlness et al. 2017 for the elastic constants
