### 2019-10-28 paudetseis/Telewavesim#11 @paudetseis
`version.txt` was deleted
