### 2019-10-28 paudetseis/Telewavesim#15 @paudetseis
A reference to Brownlee et al. (2017) was added: https://paudetseis.github.io/Telewavesim/elast.html

This paper lists all references for the elastic constants used in Telewavesim.
