### 2019-10-28 paudetseis/Telewavesim#16 @paudetseis
`telewavesim.elast.tri_tensor` was incorporated into `telewavesim.utils.set_tri_tensor` to avoid the inconsistency brought up by the reviewer.

The units of density have been clarified.
