### 2019-10-28 paudetseis/Telewavesim#18 @paudetseis
Guidelines for contributing have been added to the README on GitHub in a "Contributing" section.
