### 2019-10-28 paudetseis/Telewavesim#19 @paudetseis
The documentation for the function `utils.run_plane` has been significantly expanded with basic examples and figures: . Furthermore, these two examples (one for a "land" station and another for an "OBS" station provide examples with velocity models that incorporate general anisotropy (through their elastic tensors), also referenced in issue #21.

Other specific suggestions are welcome.
