### 2019-10-28 paudetseis/Telewavesim#21 @paudetseis
Two examples have been added in the documentation for function `utils.run_plane` (see comment in issue #19) describing usage of general anisotropy through the use of rock/mineral elastic tensors.

The notebook https://github.com/paudetseis/Telewavesim/blob/master/telewavesim/examples/Notebooks/sim_SKS.ipynb further suggests making tests with other anisotropic tensors.
