### 2019-10-28 paudetseis/Telewavesim#20 @paudetseis
All of these typos have been corrected.
