### 2019-10-28 paudetseis/Telewavesim#17 @paudetseis
The documentation for `utils.calc_ttime` has been expanded to describe typical usage of this function. An example of its usage is described in the notebook https://github.com/paudetseis/Telewavesim/blob/master/telewavesim/examples/Notebooks/sim_obs_Audet2016.ipynb.
