### 2019-10-28 #1818 @paudetseis
Thanks to @seisman and @andreww for their comments. Most of them have been addressed in new commits and comments in the corresponding issues. Issues #19 and #21 have been left open if there are other useful and specific suggestions for the documentation.
