### 2019-10-29 paudetseis/Telewavesim#23 @seisman
[JOSS review] Install from pypi
The documentation should also show how to install from pypi, i.e.:

```
pip install telewavesim
```

https://github.com/openjournals/joss-reviews/issues/1818
