### 2019-10-29 paudetseis/Telewavesim#24 @seisman
[JOSS review] Move the "How to setup the model file" manual from wiki to the documentation
This wiki page (https://github.com/paudetseis/Telewavesim/wiki/Models) is very useful for users. Why not put it in the documentation site (i.e. https://paudetseis.github.io/Telewavesim/)?

https://github.com/openjournals/joss-reviews/issues/1818
