### 2019-10-29 paudetseis/Telewavesim#25 @seisman
[JOSS review] Meaning of the column 5 if layer flag is tri
I'm not an expert on seismic anisotropy and a little confused about the meaning of column 5. I thought it's (Vmax-Vmin)/Vave. But after reading the wiki (https://github.com/paudetseis/Telewavesim/wiki/Models), and quickly going through the paper, I'm more confused. Could you add a few sentences explaining it?

https://github.com/openjournals/joss-reviews/issues/1818
