### 2019-10-29 paudetseis/Telewavesim#25 @seisman
Two related questions:

1. Does the column 5 apply to both P and S wave, i.e. P and S wave have the same anisotropy percentage?
2. How to setup general anisotropy models? Do users have to choose from the minerals predefined in the `elast` module?
