### 2019-10-29 66f9c5f
refactor utils module
* pep8
* remove note about conf module (which was removed earlier)
* replace nested loop with itertools.product
* move allowed minerals and rocks to toplevel
* fix some typos
