+++ telewavesim/utils.py
@@ -9,8 +9,8 @@
 # copies of the Software, and to permit persons to whom the Software is
 # furnished to do so, subject to the following conditions:
 
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
 
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
@@ -27,6 +27,7 @@ Utility functions to interact with ``telewavesim`` modules.
 '''
 import itertools
 import numpy as np
+from numpy import sin, cos
 from numpy.fft import fft, fftshift, ifft
 from scipy.signal import hilbert
 from obspy.core import Trace, Stream
@@ -34,6 +35,12 @@ from obspy.signal.rotate import rotate_ne_rt
 from telewavesim import elast as es
 from telewavesim.rmat_f import conf as cf_f
 from telewavesim.rmat_f import plane as pw_f
+
+
+MINERALS = ['atg', 'bt', 'cpx', 'dol', 'ep', 'grt', 'gln', 'hbl', 'jade',
+            'lws', 'lz', 'ms', 'ol', 'opx', 'plag', 'qtz', 'zo']
+
+ROCKS = ['BS_f', 'BS_m', 'EC_f', 'EC_m', 'HB', 'LHZ', 'SP_37', 'SP_80']
 
 
 def set_iso_tensor(a, b):
@@ -94,39 +101,39 @@ def set_tri_tensor(a, b, tr, pl, ani):
     NN = (b*1.e3 - db/2.)**2
     AC = (a*1.e3)**2
     FF = -LL + np.sqrt((2.*AC)**2 - 2.*AC*(AA + CC + 2.*LL) +
-            (AA + LL)*(CC + LL))
+                       (AA + LL)*(CC + LL))
     # eta = FF/(AA - 2.*LL)
 
     # Get tensor with horizontal axis
-    cc = np.zeros((3,3,3,3))
-        
-    cc[0,0,0,0] = AA
-    cc[1,1,1,1] = AA
-    cc[2,2,2,2] = CC
+    cc = np.zeros((3, 3, 3, 3))
 
-    cc[0,0,1,1] = (AA - 2.*NN)
-    cc[1,1,0,0] = (AA - 2.*NN)
+    cc[0, 0, 0, 0] = AA
+    cc[1, 1, 1, 1] = AA
+    cc[2, 2, 2, 2] = CC
 
-    cc[0,0,2,2] = FF
-    cc[2,2,0,0] = FF
+    cc[0, 0, 1, 1] = (AA - 2.*NN)
+    cc[1, 1, 0, 0] = (AA - 2.*NN)
 
-    cc[1,1,2,2] = FF
-    cc[2,2,1,1] = FF
+    cc[0, 0, 2, 2] = FF
+    cc[2, 2, 0, 0] = FF
 
-    cc[1,2,1,2] = LL
-    cc[2,1,2,1] = LL
-    cc[2,1,1,2] = LL
-    cc[1,2,2,1] = LL
+    cc[1, 1, 2, 2] = FF
+    cc[2, 2, 1, 1] = FF
 
-    cc[2,0,2,0] = LL
-    cc[0,2,0,2] = LL
[…]
