### 2019-10-29 ed5ffbc
format doc string in utils module
