### 2019-10-29 6cd8f86
add pypi install instructions, pep8
