### 2019-10-29 paudetseis/Telewavesim#21 @paudetseis
More examples on how to set up models using anisotropic tensors have been added. The Wiki page has been merged with the documentation (from a suggestion in issue #24) and there is now a section entitled "Setting up new models" with several new examples.
