### 2019-10-29 paudetseis/Telewavesim#24 @paudetseis
The Wiki page has been merged with the documentation and there is now a section entitled "Setting up new models" with several new examples.
