### 2019-10-29 paudetseis/Telewavesim#25 @paudetseis
The column 5 only applies to transversely isotropic materials, which has been clarified in the documentation. Based on issue #24 there is now a new section that describes how to setup the model from either a model file or using the Model class directly, with several new examples. There is now a note explaining what the percent anisotropy refers to, which points to the article by Porter et al., 2010.

1. Yes, the percent anisotropy applies to both Vp and Vs similarly
2. Yes, users have to choose from the list of minerals or rocks, and a note has been added to encourage users to suggest their own elastic tensor or add them to the code directly.
