### 2019-10-29 paudetseis/Telewavesim#25 @seisman
Looks good to me. Thanks.
