### 2019-10-29 #1818 @seisman
I've finished my review and the authors has addressed my issues. The submission now looks good to me.
