+++ docs/_sources/wiggles.rst.txt
@@ -1,8 +1,0 @@
-.. figure:: ../telewavesim/examples/picture/tws_logo.png
-   :align: center
-
-Module wiggles
---------------
-
-.. automodule:: telewavesim.wiggle
-   :members:
