+++ docs/_sources/utils.rst.txt
@@ -1,8 +1,0 @@
-.. figure:: ../telewavesim/examples/picture/tws_logo.png
-   :align: center
-
-Module utils
-------------
-
-.. automodule:: telewavesim.utils
-   :members:
