+++ docs/conf.py
@@ -22,7 +22,6 @@ copyright = '2019, Pascal Audet'
 author = 'Pascal Audet'
 
 # The full version, including alpha/beta/rc tags
-#release = '0.1.0'
 from telewavesim import __version__ as release
 
 # -- General configuration ---------------------------------------------------
@@ -30,14 +29,14 @@ from telewavesim import __version__ as release
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc', 'sphinx.ext.napoleon', 'sphinx.ext.githubpages']
+extensions = ['sphinx.ext.autodoc', 'sphinx.ext.napoleon']
 
 autodoc_member_order = 'bysource'
 
 html_logo = '../telewavesim/examples/picture/logo_tws.png'
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+#templates_path = ['_templates']
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -57,4 +56,4 @@ html_theme = 'sphinx_rtd_theme'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+#html_static_path = ['_static']
