### 2019-10-29 dec562c
let travis build and deploy the html doc
deploy key still needed
