### 2019-10-29 769291c
add extra job for doc builds
