### 2019-10-29 bbf71c0
try to fix .travis.yml
