### 2019-10-31 #1818 @kbarnhart
Thanks @andreww and @seisman for completing your reviews.

@andreww it looks like there are a couple of check-boxes remaining unchecked. Please post here if anything remains for @paudetseis to do for you to check them.

@brmather, please let me know if you have any questions as you undertake your review.
