### 2019-11-01 paudetseis/Telewavesim#15 @andreww
There are still a bunch of citations for the origin of elastic constants in individual functions that lack reference details / links. It would probably be best to add links/information about these too as users may well not bother to track this information back through Brownlee.
