### 2019-11-01 #1818 @andreww
Yes - I'm happy that the changes address my concerns. I've added comments to one closed issue (https://github.com/paudetseis/Telewavesim/issues/15) but I don't think this needs to be addressed to allow publication here.
