### 2019-11-04 paudetseis/Telewavesim#31 @brmather
[Review] dependency issues with installation
https://github.com/openjournals/joss-reviews/issues/1818

I am running into the following error with `pip install .` in the source directory:

```sh
    /usr/bin/ld: cannot find -llapack
    collect2: error: ld returned 1 exit status
```

which suggests that it cannot find my lapack installation. However, if I run `pip install telewavesim` then it installs fine. Is there something new in the repo that is preventing me from installing this properly?

In addition I run into the following issue when running the test suite:

```sh
tests/test_utils.py:2: in <module>
    from telewavesim import utils as ut
utils.py:33: in <module>
    from obspy.core import Trace, Stream
../../../.local/lib/python3.6/site-packages/obspy/__init__.py:40: in <module>
    from obspy.core.util import _get_version_string
../../../.local/lib/python3.6/site-packages/obspy/core/__init__.py:126: in <module>
    from obspy.core.trace import Stats, Trace  # NOQA
../../../.local/lib/python3.6/site-packages/obspy/core/trace.py:26: in <module>
    from obspy.core.util import AttribDict, create_empty_data_chunk
../../../.local/lib/python3.6/site-packages/obspy/core/util/__init__.py:27: in <module>
    from obspy.core.util.base import (ALL_MODULES, DEFAULT_MODULES,
../../../.local/lib/python3.6/site-packages/obspy/core/util/base.py:397: in <module>
    PROJ4_VERSION = get_proj_version()
../../../.local/lib/python3.6/site-packages/obspy/core/util/base.py:383: in get_proj_version
    version_string = str(Proj(proj='utm', zone=10, ellps='WGS84').proj_version)
E   AttributeError: 'Proj' object has no attribute 'proj_version'

```

which looks to be a dependency issue with `obspy` 1.1.1. Switching to conda may resolve these dependency issues, but I try to avoid Anaconda as it has the tendency to mess up my system python configuration. In light of this, I would suggest bundling a Docker image with `Telewavesim` - what do you think?
