### 2019-11-04 paudetseis/Telewavesim#31 @trichter
The package on pypi is outdated and I think it is reasonable to publish a new version only after the review. Therefore, please install from source. Are you running on Linux or OSX? The tests on travis-ci indicate it should work with recent versions of numpy both on Linux and OSX. Does it work for you if you delete the option `libraries=['lapack']` in line 18 in `setup.py`?
