+++ .travis.yml
@@ -41,31 +41,19 @@ script:
 after_success:
   - if [[ $DOCS != 1 ]]; then
       bash <(curl -s https://codecov.io/bash);
+    else
+      openssl aes-256-cbc -K $encrypted_4b1a607b6302_key -iv $encrypted_4b1a607b6302_iv -in "../.travis/github_deploy_key.enc" -out "../.travis/github_deploy_key" -d;
     fi
-  - |
-
-       declare -r SSH_FILE="$(mktemp -u $HOME/.ssh/travisdoc)"
-
-       # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
-       # Decrypt the file containing the private key
-       # (Note: this is the same as what is generated by the Travis CLI at step 2.5)
-
-       openssl aes-256-cbc \
-         -K $encrypted_4b1a607b6302_key \
-         -iv $encrypted_4b1a607b6302_iv \
-         -in ".travis/github_deploy_key.enc" \
-         -out "$SSH_FILE" -d
 
        # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
        # Enable SSH authentication
 
-       chmod 600 "$SSH_FILE" \
-         && printf "%s\n" \
-              "Host github.com" \
-              "  IdentityFile $SSH_FILE" \
-              "  LogLevel ERROR" >> ~/.ssh/config
+       #chmod 600 "$SSH_FILE" \
+       #  && printf "%s\n" \
+       #       "Host github.com" \
+       #       "  IdentityFile $SSH_FILE" \
+       #       "  LogLevel ERROR" >> ~/.ssh/config
 notifications:
     email: false
 
@@ -76,7 +64,7 @@ env:
 deploy:
   provider: pages:git
   edge: true # opt in to dpl v2
-  deploy_key: ../.travis/github_deploy_key.enc
+  deploy_key: ../.travis/github_deploy_key
   on:
     all_branches: true
     condition: $DOCS = 1
