### 2019-11-04 paudetseis/Telewavesim#31 @brmather
I'm running Linux. Deleting `libraries=['lapack']` line meant that I could install the package from source, but now I run into this error:

```python
Traceback:
telewavesim/tests/test_utils.py:2: in <module>
    from telewavesim import utils as ut
telewavesim/utils.py:36: in <module>
    from telewavesim.rmat_f import conf as cf_f
E   ModuleNotFoundError: No module named 'telewavesim.rmat_f'
```
which looks like it's not compiling the fortran source correctly? The second issue I was having is not related to `Telewavesim` but rather a conflict in `obspy` with a later version of `pyproj` (https://github.com/obspy/obspy/issues/2445 and https://github.com/obspy/obspy/issues/2449).
