+++ wiggles.rst
@@ -1,0 +1,8 @@
+.. figure:: ../telewavesim/examples/picture/tws_logo.png
+   :align: center
+
+Module wiggles
+--------------
+
+.. automodule:: telewavesim.wiggle
+   :members:
