### 2019-11-04 4827cea
clean up .tavis.yml, add local_dir option
