+++ .travis.yml
@@ -36,24 +36,15 @@ script:
   - if [[ $DOCS != 1 ]]; then
       mkdir empty; cd empty; pytest -v --cov=telewavesim --pyargs telewavesim;
     else
-      cd docs; make html;
+      cd docs; make html; cd ..;
     fi
 after_success:
   - if [[ $DOCS != 1 ]]; then
       bash <(curl -s https://codecov.io/bash);
     else
-      openssl aes-256-cbc -K $encrypted_4b1a607b6302_key -iv $encrypted_4b1a607b6302_iv -in "../.travis/github_deploy_key.enc" -out "../.travis/github_deploy_key" -d;
+      openssl aes-256-cbc -K $encrypted_4b1a607b6302_key -iv $encrypted_4b1a607b6302_iv -in ".travis/github_deploy_key.enc" -out ".travis/github_deploy_key" -d;
     fi
 
-       # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
-       # Enable SSH authentication
-
-       #chmod 600 "$SSH_FILE" \
-       #  && printf "%s\n" \
-       #       "Host github.com" \
-       #       "  IdentityFile $SSH_FILE" \
-       #       "  LogLevel ERROR" >> ~/.ssh/config
 notifications:
     email: false
 
@@ -64,7 +55,8 @@ env:
 deploy:
   provider: pages:git
   edge: true # opt in to dpl v2
-  deploy_key: ../.travis/github_deploy_key
+  deploy_key: .travis/github_deploy_key
+  local_dir: docs/_build/html
   on:
     all_branches: true
     condition: $DOCS = 1
