### 2019-11-04 f840bef
check if it is working without env global secure
