### 2019-11-04 paudetseis/Telewavesim#31 @trichter
It would not have surprised me, if you ran into `ImpoertError - undefined symbol: dgeev_`, but this error seems strange. Are you still in the source directory? Maybe it is trying to import from there. Can you try to run the tests from an empty folder?
