### 2019-11-04 paudetseis/Telewavesim#31 @trichter
> In light of this, I would suggest bundling a Docker image with Telewavesim - what do you think?

In my opinion, this would be overkill. Hopefully, in the end, there will be package on pypi for which the installation is simply working out of the box.
