### 2019-11-04 f2819bb
added touch .nojekyll command to travis.yml
