### 2019-11-05 paudetseis/Telewavesim#31 @brmather
I have compiled lapack from source with shared libraries using cmake. When I run
```
f2py -c *.f90 -L/usr/local/lib64 -llapack -m rmat_f
```
from the `src` folder it seems to compile fine but it returns an `ImportError`:

```
ImportError                               Traceback (most recent call last)
<ipython-input-1-553ea2ea12ab> in <module>
----> 1 import rmat_f

ImportError: liblapack.so.3: cannot open shared object file: No such file or directory
```

`/usr/local/lib64/liblapack.so.3` _is_ there, but somehow f2py fails to link to it.

```
$ ldd rmat_f.cpython-36m-x86_64-linux-gnu.so
	linux-vdso.so.1 (0x00007ffeb1569000)
	liblapack.so.3 => not found
	libpython3.6m.so.1.0 => /usr/lib/libpython3.6m.so.1.0 (0x00007f3987c75000)
	libgfortran.so.5 => /usr/lib/libgfortran.so.5 (0x00007f3987668000)
	libm.so.6 => /usr/lib/haswell/libm.so.6 (0x00007f3987528000)
	libgcc_s.so.1 => /usr/lib/libgcc_s.so.1 (0x00007f398750d000)
	libquadmath.so.0 => /usr/lib/libquadmath.so.0 (0x00007f39874bd000)
	libc.so.6 => /usr/lib/haswell/libc.so.6 (0x00007f39872d4000)
	libpthread.so.0 => /usr/lib/libpthread.so.0 (0x00007f39872b3000)
	libdl.so.2 => /usr/lib/libdl.so.2 (0x00007f39872ad000)
	libutil.so.1 => /usr/lib/libutil.so.1 (0x00007f39872a8000)
	/usr/lib64/ld-linux-x86-64.so.2 (0x00007f3988030000)

```

Any advice on how to proceed would be appreciated.
