### 2019-11-05 paudetseis/Telewavesim#31 @trichter
I also do not know, what's wrong, but I have some ideas how to proceed:

* set `export NPY_DISTUTILS_APPEND_FLAGS=1` before calling `f2py`, this will be default in future numpy versions and we also set it when installing the package on travis-ci.
* What is your numpy version? Is it `>=1.15`? Some installation errors went away on travis, after pinning numpy to `>=1.15`.
* Can you post the output of the f2py command?
* For me it was sometimes even necessary to open a new terminal, but that was probably due to conda.
