### 2019-11-05 paudetseis/Telewavesim#31 @brmather
- `export NPY_DISTUTILS_APPEND_FLAGS=1` doesn't appear to change anything.
- My numpy version is 1.17.3
- Here is the full output of `f2py`.
```
running build
running config_cc
unifing config_cc, config, build_clib, build_ext, build commands --compiler options
running config_fc
unifing config_fc, config, build_clib, build_ext, build commands --fcompiler options
running build_src
build_src
building extension "rmat_f" sources
f2py options: []
f2py:> /tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/rmat_fmodule.c
creating /tmp/tmpiwu4ku4f/src.linux-x86_64-3.6
Reading fortran codes...
	Reading file 'rmat.f90' (format:free)
	Reading file 'rmat_sub.f90' (format:fix)
Post-processing...
	Block: rmat_f
			Block: conf
In: :rmat_f:rmat.f90:conf
get_parameters: got "unexpected EOF while parsing (<string>, line 1)" on '3.141592653589793d0'
			Block: rmat
				Block: rmatrix
In: :rmat_f:rmat.f90:rmat:rmatrix
get_parameters: got "unexpected EOF while parsing (<string>, line 1)" on '3.141592653589793d0'
				Block: rtfluid
In: :rmat_f:rmat.f90:rmat:rtfluid
get_parameters: got "unexpected EOF while parsing (<string>, line 1)" on '3.141592653589793d0'
			Block: plane
				Block: plane_land
In: :rmat_f:rmat.f90:plane:plane_land
get_parameters: got "unexpected EOF while parsing (<string>, line 1)" on '3.141592653589793d0'
				Block: plane_obs
In: :rmat_f:rmat.f90:plane:plane_obs
get_parameters: got "unexpected EOF while parsing (<string>, line 1)" on '3.141592653589793d0'
			Block: layer
			Block: addit
			Block: xeveci
			Block: isotroc
			Block: anisotroc
			Block: xinv3
			Block: rxinv3
			Block: xmult3
			Block: sort_evec
			Block: reorder_evec
			Block: sort_r
			Block: cnorm
Post-processing (stage 2)...
	Block: rmat_f
		Block: unknown_interface
			Block: conf
			Block: rmat
				Block: rmatrix
				Block: rtfluid
			Block: plane
				Block: plane_land
				Block: plane_obs
			Block: layer
			Block: addit
			Block: xeveci
			Block: isotroc
			Block: anisotroc
			Block: xinv3
			Block: rxinv3
			Block: xmult3
			Block: sort_evec
			Block: reorder_evec
			Block: sort_r
			Block: cnorm
Building modules...
	Building module "rmat_f"...
		Constructing wrapper function "layer"...
		  layer(omega,xh,eval,tu,td,rd)
		Constructing wrapper function "addit"...
		  addit(nlay,ilay,tus,rus,tds,rds,tu,ru,td,rd)
		Constructing wrapper function "xeveci"...
		  xeveci(evec,eveci)
		Constructing wrapper function "isotroc"...
		  isotroc(vp,vs,dens,p11,p22,q,n)
		Constructing wrapper function "anisotroc"...
		  anisotroc(c,dens,p11,p22,q,n)
		Constructing wrapper function "xinv3"...
		  xinv3(a,ainv)
		Constructing wrapper function "rxinv3"...
		  rxinv3(a,ainv)
		Constructing wrapper function "xmult3"...
		  xmult3(a,b,c)
		Constructing wrapper function "sort_evec"...
		  sort_evec(q,evec,[n])
		Constructing wrapper function "reorder_evec"...
		  reorder_evec(evalr,evali,evecr,eveci,indx,eval,evec,[n])
		Constructing wrapper function "sort_r"...
		  sort_r(a,indx,[m,n])
		Constructing wrapper function "cnorm"...
		  cnorm(n,[nm])
		Constructing F90 module support for "conf"...
		  Variables: pi nlaymx a thickn rho dp c rhof isoflg dt slow baz evals evecs tui rui tdi rdi
		Constructing F90 module support for "rmat"...
			Constructing wrapper function "rmat.rmatrix"...
			  tus,rus,tds,rds = rmatrix(px,py,omega,nlay,iter)
			Constructing wrapper function "rmat.rtfluid"...
			  rtmat = rtfluid(p1,p2,a0,b0,rhos)
		Constructing F90 module support for "plane"...
			Constructing wrapper function "plane.plane_land"...
			  yx,yy,yz = plane_land(nt,nlay,wvtype)
			Constructing wrapper function "plane.plane_obs"...
			  yx,yy,yz = plane_obs(nt,nlay,wvtype)
	Wrote C/API module "rmat_f" to file "/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/rmat_fmodule.c"
	Fortran 90 wrappers are saved to "/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/rmat_f-f2pywrappers2.f90"
  adding '/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/fortranobject.c' to sources.
  adding '/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6' to include_dirs.
copying /home/ben/.local/lib/python3.6/site-packages/numpy/f2py/src/fortranobject.c -> /tmp/tmpiwu4ku4f/src.linux-x86_64-3.6
copying /home/ben/.local/lib/python3.6/site-packages/numpy/f2py/src/fortranobject.h -> /tmp/tmpiwu4ku4f/src.linux-x86_64-3.6
  adding '/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/rmat_f-f2pywrappers2.f90' to sources.
build_src: building npy-pkg config files
running build_ext
customize UnixCCompiler
customize UnixCCompiler using build_ext
get_default_fcompiler: matching types: '['gnu95', 'intel', 'lahey', 'pg', 'absoft', 'nag', 'vast', 'compaq', 'intele', 'intelem', 'gnu', 'g95', 'pathf95', 'nagfor']'
customize Gnu95FCompiler
Found executable /usr/bin/gfortran
customize Gnu95FCompiler
customize Gnu95FCompiler using build_ext
building 'rmat_f' extension
compiling C sources
C compiler: x86_64-solus-linux-gcc -pthread -Wno-unused-result -Wsign-compare -DNDEBUG -g -fwrapv -O3 -Wall -mtune=generic -march=x86-64 -g2 -O2 -pipe -fPIC -Wformat -Wformat-security -D_FORTIFY_SOURCE=2 -fstack-protector-strong --param ssp-buffer-size=32 -fasynchronous-unwind-tables -ftree-vectorize -feliminate-unused-debug-types -Wall -Wno-error -Wp,-D_REENTRANT -I/usr/include -mtune=generic -march=x86-64 -g2 -O2 -pipe -fPIC -Wformat -Wformat-security -D_FORTIFY_SOURCE=2 -fstack-protector-strong --param ssp-buffer-size=32 -fasynchronous-unwind-tables -ftree-vectorize -feliminate-unused-debug-types -Wall -Wno-error -Wp,-D_REENTRANT -fPIC

creating /tmp/tmpiwu4ku4f/tmp
creating /tmp/tmpiwu4ku4f/tmp/tmpiwu4ku4f
creating /tmp/tmpiwu4ku4f/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6
compile options: '-I/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6 -I/home/ben/.local/lib/python3.6/site-packages/numpy/core/include -I/usr/include/python3.6m -c'
x86_64-solus-linux-gcc: /tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/fortranobject.c
x86_64-solus-linux-gcc: /tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/rmat_fmodule.c
In file included from /home/ben/.local/lib/python3.6/site-packages/numpy/core/include/numpy/ndarraytypes.h:1830,
                 from /home/ben/.local/lib/python3.6/site-packages/numpy/core/include/numpy/ndarrayobject.h:12,
                 from /home/ben/.local/lib/python3.6/site-packages/numpy/core/include/numpy/arrayobject.h:4,
                 from /tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/fortranobject.h:13,
                 from /tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/fortranobject.c:2:
/home/ben/.local/lib/python3.6/site-packages/numpy/core/include/numpy/npy_1_7_deprecated_api.h:17:2: warning: #warning "Using deprecated NumPy API, disable it with " "#define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION" [-Wcpp]
   17 | #warning "Using deprecated NumPy API, disable it with " \
      |  ^~~~~~~
In file included from /home/ben/.local/lib/python3.6/site-packages/numpy/core/include/numpy/ndarraytypes.h:1830,
                 from /home/ben/.local/lib/python3.6/site-packages/numpy/core/include/numpy/ndarrayobject.h:12,
                 from /home/ben/.local/lib/python3.6/site-packages/numpy/core/include/numpy/arrayobject.h:4,
                 from /tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/fortranobject.h:13,
                 from /tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/rmat_fmodule.c:16:
/home/ben/.local/lib/python3.6/site-packages/numpy/core/include/numpy/npy_1_7_deprecated_api.h:17:2: warning: #warning "Using deprecated NumPy API, disable it with " "#define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION" [-Wcpp]
   17 | #warning "Using deprecated NumPy API, disable it with " \
      |  ^~~~~~~
/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/rmat_fmodule.c:144:12: warning: ‘f2py_size’ defined but not used [-Wunused-function]
  144 | static int f2py_size(PyArrayObject* var, ...)
      |            ^~~~~~~~~
compiling Fortran 90 module sources
Fortran f77 compiler: /usr/bin/gfortran -Wall -g -ffixed-form -fno-second-underscore -fPIC -O3 -funroll-loops
Fortran f90 compiler: /usr/bin/gfortran -Wall -g -fno-second-underscore -fPIC -O3 -funroll-loops
Fortran fix compiler: /usr/bin/gfortran -Wall -g -ffixed-form -fno-second-underscore -Wall -g -fno-second-underscore -fPIC -O3 -funroll-loops
compile options: '-I/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6 -I/home/ben/.local/lib/python3.6/site-packages/numpy/core/include -I/usr/include/python3.6m -c'
extra options: '-J/tmp/tmpiwu4ku4f/ -I/tmp/tmpiwu4ku4f/'
gfortran:f90: rmat.f90
compiling Fortran sources
Fortran f77 compiler: /usr/bin/gfortran -Wall -g -ffixed-form -fno-second-underscore -fPIC -O3 -funroll-loops
Fortran f90 compiler: /usr/bin/gfortran -Wall -g -fno-second-underscore -fPIC -O3 -funroll-loops
Fortran fix compiler: /usr/bin/gfortran -Wall -g -ffixed-form -fno-second-underscore -Wall -g -fno-second-underscore -fPIC -O3 -funroll-loops
compile options: '-I/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6 -I/home/ben/.local/lib/python3.6/site-packages/numpy/core/include -I/usr/include/python3.6m -c'
extra options: '-J/tmp/tmpiwu4ku4f/ -I/tmp/tmpiwu4ku4f/'
gfortran:f90: rmat_sub.f90
gfortran:f90: /tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/rmat_f-f2pywrappers2.f90
/usr/bin/gfortran -Wall -g -Wall -g -shared /tmp/tmpiwu4ku4f/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/rmat_fmodule.o /tmp/tmpiwu4ku4f/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/fortranobject.o /tmp/tmpiwu4ku4f/rmat.o /tmp/tmpiwu4ku4f/rmat_sub.o /tmp/tmpiwu4ku4f/tmp/tmpiwu4ku4f/src.linux-x86_64-3.6/rmat_f-f2pywrappers2.o -L/usr/local/lib64 -L/usr/lib64/gcc/x86_64-solus-linux/9/../../../../lib64 -L/usr/lib64/gcc/x86_64-solus-linux/9/../../../../lib64 -L/usr/lib64 -llapack -lpython3.6m -lgfortran -o ./rmat_f.cpython-36m-x86_64-linux-gnu.so
Removing build directory /tmp/tmpiwu4ku4f

```
