### 2019-11-06 paudetseis/Telewavesim#31 @trichter
I am not an expert in this linking stuff. Nevertheless, maybe answers to this [stackoverflow question](https://stackoverflow.com/questions/480764/linux-error-while-loading-shared-libraries-cannot-open-shared-object-file-no-s) are helpful?

Does `ldconfig -p | grep lapack` output your freshly compiled library?
Does `cat /etc/ld.so.conf.d/*.conf | grep /usr/local/lib64` output your lib directory?

If not that might be the reason for the problems. I think you could solve this problem by adding a new file to the `/etc/ld.so.conf.d` folder with `sudo echo "/usr/local/lib64" > /etc/ld.so.conf.d/usrlocallib64.conf` running `sudo ldconfig` afterwards. (Alternatively check and set your $LD_LIBRARY_PATH environment variable).
