### 2019-11-07 paudetseis/Telewavesim#31 @brmather
That worked! I had to manually add `/usr/local/lib64/` to my library path for `rmat_f` to link against. Also, the fix in https://github.com/paudetseis/Telewavesim/issues/2#issuecomment-538487241 was required otherwise the fortran sources do not detect lapack (even if it's in my library path). So long as other installation methods work with these changes to `setup.py`, they should be commited to `master`.
