### 2019-11-07 08dd02f
refactor .travis.yml
