+++ .travis.yml
@@ -1,33 +1,27 @@
 language: c
 sudo: false
-matrix:
+os:
+  - linux
+  - osx
+env:
+  - PYTHON_VERSION=3.5
+  - PYTHON_VERSION=3.6
+  - PYTHON_VERSION=3.7
+jobs:
   include:
-    - os: linux
-      env: PYTHON_VERSION=3.5
-    - os: linux
-      env: PYTHON_VERSION=3.6
-    - os: linux
-      env: PYTHON_VERSION=3.7
-    - os: linux
-      env: PYTHON_VERSION=3.7 DOCS=1
-    - os: osx
-      env: PYTHON_VERSION=3.5
-    - os: osx
-      env: PYTHON_VERSION=3.6
-    - os: osx
-      env: PYTHON_VERSION=3.7
-
+    - env: PYTHON_VERSION=3.7 DOCS=1
 before_install:
   - export NPY_DISTUTILS_APPEND_FLAGS=1
   - wget https://raw.githubusercontent.com/trichter/conda4travis/latest/conda4travis.sh -O conda4travis.sh
   - source conda4travis.sh
   - conda config --add channels conda-forge
-  - if [[ $DOCS == 1 ]]; then
+  - if [[ $DOCS != 1 ]]; then
+      DEP="pytest-cov";
+    else
       DEP="sphinx sphinx_rtd_theme";
     fi
   - conda create -q -n testenv
-      python=$PYTHON_VERSION obspy "gfortran_${TRAVIS_OS_NAME}-64" "numpy>=1.15"
-      pytest-cov $DEP
+      python=$PYTHON_VERSION obspy "gfortran_${TRAVIS_OS_NAME}-64" "numpy>=1.15" $DEP
   - conda activate testenv
   - conda list
 install:
