### 2019-11-07 5d8dea8
test all numpy versions >= 1.15
