+++ .travis.yml
@@ -4,9 +4,15 @@ os:
   - linux
   - osx
 env:
-  - PYTHON_VERSION=3.5
-  - PYTHON_VERSION=3.6
-  - PYTHON_VERSION=3.7
+  - PYTHON_VERSION=3.5 NPY="=1.15"
+  - PYTHON_VERSION=3.5 NPY="=1.16"
+  - PYTHON_VERSION=3.5 NPY="=1.17"
+  - PYTHON_VERSION=3.6 NPY="=1.15"
+  - PYTHON_VERSION=3.6 NPY="=1.16"
+  - PYTHON_VERSION=3.6 NPY="=1.17"
+  - PYTHON_VERSION=3.7 NPY="=1.15"
+  - PYTHON_VERSION=3.7 NPY="=1.16"
+  - PYTHON_VERSION=3.7 NPY="=1.17"
 jobs:
   include:
     - env: PYTHON_VERSION=3.7 DOCS=1
@@ -21,7 +27,7 @@ before_install:
       DEP="sphinx sphinx_rtd_theme";
     fi
   - conda create -q -n testenv
-      python=$PYTHON_VERSION obspy "gfortran_${TRAVIS_OS_NAME}-64" "numpy>=1.15" $DEP
+      python=$PYTHON_VERSION obspy "gfortran_${TRAVIS_OS_NAME}-64" "numpy$NPY" $DEP
   - conda activate testenv
   - conda list
 install:
