### 2019-11-07 ccfeabc
test also with gfortran from system
