+++ .travis.yml
@@ -1,21 +1,17 @@
 language: c
-sudo: false
 os:
   - linux
   - osx
 env:
-  - PYTHON_VERSION=3.5 NPY="=1.15"
-  - PYTHON_VERSION=3.5 NPY="=1.16"
-  - PYTHON_VERSION=3.5 NPY="=1.17"
-  - PYTHON_VERSION=3.6 NPY="=1.15"
-  - PYTHON_VERSION=3.6 NPY="=1.16"
-  - PYTHON_VERSION=3.6 NPY="=1.17"
-  - PYTHON_VERSION=3.7 NPY="=1.15"
-  - PYTHON_VERSION=3.7 NPY="=1.16"
-  - PYTHON_VERSION=3.7 NPY="=1.17"
+  - PYTHON_VERSION=3.5 NPY_PIN_VERSION="=1.15"
+  - PYTHON_VERSION=3.5
+  - PYTHON_VERSION=3.6
+  - PYTHON_VERSION=3.7
 jobs:
   include:
     - env: PYTHON_VERSION=3.7 DOCS=1
+    - env: PYTHON_VERSION=3.7 NPY_PIN_VERSION=">=1.17" SYSTEMGFORTRAN=1
+      sudo: required
 before_install:
   - export NPY_DISTUTILS_APPEND_FLAGS=1
   - wget https://raw.githubusercontent.com/trichter/conda4travis/latest/conda4travis.sh -O conda4travis.sh
@@ -26,8 +22,13 @@ before_install:
     else
       DEP="sphinx sphinx_rtd_theme";
     fi
+  - if [[ $SYSTEMGFORTRAN != 1 ]]; then
+      GFORTRAN="gfortran_${TRAVIS_OS_NAME}-64";
+    else
+      sudo apt-get install gfortran;
+    fi
   - conda create -q -n testenv
-      python=$PYTHON_VERSION obspy "gfortran_${TRAVIS_OS_NAME}-64" "numpy$NPY" $DEP
+      python=$PYTHON_VERSION obspy $GFORTRAN "numpy$NPY_PIN_VERSION" $DEP
   - conda activate testenv
   - conda list
 install:
