### 2019-11-07 95a7f05
add lapack work-around and use library_dirs option
see #2 and #31
