### 2019-11-07 paudetseis/Telewavesim#31 @trichter
I pushed a similar fix to master. I would appreciate if you could test it with your setup.
