### 2019-11-13 paudetseis/Telewavesim#31 @paudetseis
Thanks @trichter . @brmather , did you have a chance to test the fix? Hopefully it's all good now and we can move on.
