### 2019-11-14 e2e1479
Automatically reference dependencies on import
