+++ telewavesim/__init__.py
@@ -404,5 +404,26 @@ Single event for OBS station
 
 
 """
+from numpy.distutils.system_info import get_info as _get_info
+import os as _os
+import sys as _sys
+from ctypes import cdll as _cdll
+
+try:
+    from . import rmat_f
+except:
+
+    _shared_libraries = []
+
+    for path in _get_info('lapack')['library_dirs']:
+        for lib in _os.listdir(path):
+            if lib.startswith('lib'):
+                _slib = _cdll.LoadLibrary("{}/{}".format(path, lib))
+                _shared_libraries.append( _slib )
+
+        # _sys.path.insert(0, str(path))
+        # _os.environ['LD_LIBRARY_PATH'] += ":"+str(path)
+    # _os.execv(_sys.argv[0], _sys.argv)
+
 
 __version__ = '0.2.0-dev'
