### 2019-11-14 f0919c8
remove a few commented out bits
