### 2019-11-14 paudetseis/Telewavesim#31 @brmather
Hi @trichter and @paudetseis,

The workaround in #33 did not work for me. The function `get_library_dirs` in `setup.py` could not locate lapack on my system so I submitted PR #34. This re-implements numpy's `get_info` function to link lapack on compiling the fortran source code, and ensures the shared libraries are loaded on importing `telewavesim`. I found [this forum thread](https://stackoverflow.com/questions/6543847/setting-ld-library-path-from-inside-python) quite helpful.

This will likely not be an issue for the majority of users on Anaconda, but it should harden the installation procedure for people like me on native installs.

Feel free to modify the pull request as you see fit. As it stands I'm happy to close this issue and move on.
