+++ telewavesim/__init__.py
@@ -172,20 +172,20 @@ Setting up new models
 From the model file
 ~~~~~~~~~~~~~~~~~~~
 
-In the Jupiter notebooks you will find a folder named ``models`` where a 
-few examples are provided. The header of the file ``model_Audet2016.txt`` 
+In the Jupiter notebooks you will find a folder named ``models`` where a
+few examples are provided. The header of the file ``model_Audet2016.txt``
 looks like:
 
 .. sourcecode:: bash
 
    ################################################
-   # 
-   #   Model file to use with `telewavesim` for 
-   #   modeling teleseismic body wave propagation 
+   #
+   #   Model file to use with `telewavesim` for
+   #   modeling teleseismic body wave propagation
    #   through stratified media.
    #
-   #   Lines starting with '#' are ignored. Each 
-   #   line corresponds to a unique layer. The 
+   #   Lines starting with '#' are ignored. Each
+   #   line corresponds to a unique layer. The
    #   bottom layer is assumed to be a half-space
    #   (Thickness is irrelevant).
    #
@@ -195,7 +195,7 @@ looks like:
    #          1    Density (kg/m^3)
    #          2    Layer P-wave velocity (km/s)
    #          3    Layer S-wave velocity (km/s)
-   #          4    Layer flag 
+   #          4    Layer flag
    #                   iso: isotropic
    #                   tri: transverse isotropy
    #                   [other]: other minerals or rocks
@@ -208,9 +208,9 @@ looks like:
    #
    ################################################
 
-The header is not required and can be deleted when you become familiar 
-with the various definitions. Note that the code requires 8 entries per 
-layer, regardless of whether or not the variable is required (it will 
+The header is not required and can be deleted when you become familiar
+with the various definitions. Note that the code requires 8 entries per
+layer, regardless of whether or not the variable is required (it will
 simply be ignored if it's not).
 
 Let us break down each line, depending on how you set ``Layer flag``:
@@ -227,15 +227,15 @@ This flag represents a case where the layer is isotropic.
 Layer flag set to ``tri``
 *************************
 
-This flag represents a transversely isotropic layer. We adhere with 
-the definition in 
-`Porter et al. (2011) <https://doi.org/10.1130/L126.1>`_, 
-whereby the parameter ``eta``, which describes the curvature of the 
-velocity “ellipsoid” between the Vp-fast and Vp-slow axes, varies 
-with anisotropy for a 2-psi model and is not fixed. 
+This flag represents a transversely isotropic layer. We adhere with
+the definition in
+`Porter et al. (2011) <https://doi.org/10.1130/L126.1>`_,
+whereby the parameter ``eta``, which describes the curvature of the
+velocity “ellipsoid” between the Vp-fast and Vp-slow axes, varies
+with anisotropy for a 2-psi model and is not fixed.
 
 The column 5 in this case sets the percent anisotropy for both
-Vp and Vs (equal anisotropy for both Vp and Vs) and is the only 
+Vp and Vs (equal anisotropy for both Vp and Vs) and is the only
 instance where this column is required.
 
 - Set all columns to the required numerical value (and column 4 to ``tri``)
@@ -243,8 +243,8 @@ instance where this column is required.
 Layer flag set to another type of material
 ******************************************
 
-This flag should be set to the specific single-crystal or rock 
-abbreviation, for which the elastic properties have been determined 
+This flag should be set to the specific single-crystal or rock
+abbreviation, for which the elastic properties have been determined
 in the lab. Currently available options are:
[…]
