### 2019-11-15 2d4c36a
move imports into except block
