### 2019-11-15 a056b4c
do not pin numpy for system gfortran
