### 2019-11-19 #1818 @kbarnhart
I wanted to check in regarding this review.
