### 2019-11-20 #1818 @brmather
Hi all, I have progressed through the checklist and I find all the checklist items have been met. Aside from some installation issues, that have been rectified, I did not have any additional issues to raise because they were addressed by other reviewers. The API documentation is particularly rich :clap: As it stands, I consider `Telewavesim` to be a valuable tool for the community and ready for publication.

Thanks for the opportunity to provide a review and congrats to the authors :+1:
