### 2019-11-21 #1818 @kbarnhart
@seisman @brmather @andreww thank you all for your reviews. As part of the JOSS review process, can you please confirm your recommendation to accept the submission.
