### 2019-11-22 #1818 @andreww
Noting that I've not looked at any changes that may have been made in the last 27 days, my recommendation is that this manuscript should be accepted by JOSS.
