### 2019-11-22 #1818 @kbarnhart
Thank you @seisman @brmather @andreww

@paudetseis at this point could you:
- [ ] Edit the paper to clarify affiliation 3. It looks like two affiliations.
- [ ] Archive the reviewed software in Zenodo or a  similar service.
- [ ] Verify that the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
- [ ] Please list the Zenodo DOI of the archived version here.
