+++ telewavesim/__init__.py
@@ -421,4 +421,4 @@ except:
                 _shared_libraries.append( _slib )
 
 
-__version__ = '0.2.0-dev'
+__version__ = '0.2.0'
