### 2019-11-23 #1818 @paudetseis
@kbarnhart @seisman @brmather @andreww

Thank you all very much for taking time to review and improve Telewavesim! This was a great learning experience for me and I appreciate the valuable comments and suggestions.

I have completed the tasks listed above.

- I edited the paper and fixed affiliation 3
- The version tag of the archived version is 0.2.0
- The package on PyPi was also updated to 0.2.0
- I edited the software details on Zenodo.
- The Zenodo deposit does have the correct metadata (title + author list).
- The Zenodo DOI is 10.5281/zenodo.3551250 (also shown as badge on the README page in the telewavesim repo).

I haven't recompiled the PDF and will leave that to @kbarnhart once these details are verified and confirmed.
