### 2019-11-23 #1818 @kbarnhart
@paudetseis I found that the DOI you mentioned here is not resolved (https://doi.org/10.5281/zenodo.203551250).

Following the link on the readme I find: http://doi.org/10.5281/zenodo.3551258

Can you confirm the DOI?
