### 2019-11-24 #1818 @paudetseis
@kbarnhart I am confused about the DOI that you cite (https://doi.org/10.5281/zenodo.203551250), as I have not mentioned that DOI previously nor can I find any reference to it on this page, in the repo, README or API. The one mentioned above is the correct one (10.5281/zenodo.3551250),which is the same that appears on the README.

Perhaps I misunderstood your final check box about the zenodo DOI: "Please list the Zenodo DOI of the archived version here." Can you please clarify where you found reference to the unresolved DOI for Telewavesim and how I can fix this?
