### 2019-11-24 #1818 @kbarnhart
@paudetseis My apologies. I had meant to type https://doi.org/10.5281/zenodo.3551250
instead of https://doi.org/10.5281/zenodo.203551250. I mistyped and added a "20".

However, I do not find the DO https://doi.org/10.5281/zenodo.3551250 as resolved.

When I try to go to it I get the following screen.

I then went to the Telewavesim repository and the badge there lists:
10.5281/zenodo.3551258

So I wanted to verify the correct DOI. I suspect it should be 10.5281/zenodo.3551258

(FYI - I'll be away from internet for 4 days starting mid-day Monday Nov 25th in GMT+12)
