### 2019-11-28 #1818 @kbarnhart
@openjournals/joss-eics this is ready for final processing.
