### 2019-12-01 #1818 @Kevin-Mattheus-Moerman
@paudetseis can you please work on the following:

- [ ] Can you amend the meta data for the archive on Zenodo such that the title and authors match the JOSS paper?
- [ ] The paper contains two instances of `allows`, I suggest using `allows the` instead.
- [ ] The acronym OBS is defined in the text but never used. Perhaps you can omit the use of the acronym altogether.
- [ ] In the following: `...from core-refracted teleseismic shear waves (i.e., SKS, SKKS), etc.`, please consider removing `, etc.`, at the end of the sentence. Since the sentence contains `such as`, it is already clear that what follows is a subset of non-exhaustive examples.
