### 2019-12-02 #1818 @paudetseis
@Kevin-Mattheus-Moerman Thank you for the suggestions.

- I edited the title and authors on Zenodo to match the JOSS paper.
- The two instances of `allows` are not followed by `the`  but rather by a noun and a gerund, respectively, which I believe is correct grammar.
- I removed the acronym OBS.
- I amended the sentence and removed `, etc.` at the end of the sentence.
